+++ shuup/core/models/__init__.py
@@ -87,7 +87,7 @@ from ._service_shipping import Carrier, CustomCarrier, ShippingMethod
 from ._shipments import Shipment, ShipmentProduct, ShipmentStatus, ShipmentType
 from ._shops import Shop, ShopStatus
 from ._supplied_products import SuppliedProduct
-from ._suppliers import Supplier, SupplierShop, SupplierType
+from ._suppliers import Supplier, SupplierModule, SupplierShop, SupplierType
 from ._taxes import CustomerTaxGroup, Tax, TaxClass
 from ._units import DisplayUnit, PiecesSalesUnit, SalesUnit, UnitInterface
 
@@ -159,7 +159,6 @@ __all__ = [
     "PolymorphicShuupModel",
     "PolymorphicTranslatableShuupModel",
     "Product",
-    "Product",
     "ProductAttribute",
     "ProductCrossSell",
     "ProductCrossSellType",
@@ -201,6 +200,7 @@ __all__ = [
     "Supplier",
     "SupplierShop",
     "SupplierType",
+    "SupplierModule",
     "Tax",
     "TaxClass",
     "TranslatableShuupModel",