+++ shuup/core/modules/interface.py
@@ -8,9 +8,8 @@
 from __future__ import unicode_literals, with_statement
 
 import six
-from django.utils.translation import ugettext_lazy as _
 
-from shuup.apps.provides import get_provide_objects, get_provide_specs_and_objects
+from shuup.apps.provides import get_provide_specs_and_objects
 from shuup.utils.importing import load
 from shuup.utils.text import force_ascii
 
@@ -20,45 +19,32 @@ class ModuleNotFound(ValueError):
 
 
 class ModuleInterface(object):
-    _cached_module_impl = None
-    default_module_spec = None  # Overridden on class level
-    module_identifier = None  # Overridden on class level
+    _cached_modules_impl = None
     module_options_field = "module_data"  # May be overridden on class level
     module_provides_key = None
 
-    def _load_module(self):
-        spec = self.default_module_spec
-        module_identifier = self.module_identifier
-        if module_identifier:
-            impls = self.get_module_implementation_map()
-            if module_identifier not in impls:
-                raise ModuleNotFound(
-                    "Invalid module identifier %r in %s" % (module_identifier, force_ascii(repr(self)))
-                )
-            spec = impls[module_identifier]
-
-        cls = load(spec, context_explanation="Loading module for %s" % force_ascii(repr(self)))
-
+    def _load_modules(self):
+        enabled_supplier_modules = self.supplier_modules.all()
+        loaded_modules = []
         options = getattr(self, self.module_options_field, None) or {}
-        return cls(self, options)
+
+        for supplier_module in enabled_supplier_modules:
+            impls = self.get_module_implementation_map()
+            if supplier_module.module_identifier not in impls:
+                raise ModuleNotFound(
+                    "Invalid module identifier %r in %s" % (supplier_module.name, force_ascii(repr(self)))
+                )
+            spec = impls[supplier_module.module_identifier]
+            module = load(spec, context_explanation="Loading module for %s" % force_ascii(repr(self)))
+            loaded_modules.append(module(self, options))
+
+        return loaded_modules
 
     @property
-    def module(self):
-        if not getattr(self, "_cached_module_impl", None):
-            self._cached_module_impl = self._load_module()
-        return self._cached_module_impl
-
-    @classmethod
-    def get_module_choices(cls, empty_label=None):
-        if empty_label is None:
-            empty_label = _("No Choice")
-        choices = [("", empty_label)]
-
-        for module in get_provide_objects(cls.module_provides_key):
-            if module.identifier:
-                choices.append((module.identifier, getattr(module, "name", None) or module.identifier))
-        choices.sort()
-        return choices
+    def modules(self):
+        if not getattr(self, "_cached_modules_impl", None):
+            self._cached_modules_impl = self._load_modules()
+        return self._cached_modules_impl
 
     @classmethod
     def get_module_implementation_map(cls):