+++ shuup/core/signal_handlers.py
@@ -6,7 +6,7 @@
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
 from django.db.backends.signals import connection_created
-from django.db.models.signals import m2m_changed, post_save
+from django.db.models.signals import m2m_changed, post_migrate, post_save
 from django.dispatch import receiver
 
 from shuup.core.models import (
@@ -35,6 +35,13 @@ from shuup.core.utils.context_cache import (
 )
 from shuup.core.utils.db import extend_sqlite_functions
 from shuup.core.utils.price_cache import bump_all_price_caches, bump_prices_for_product, bump_prices_for_shop_product
+
+
+@receiver(post_migrate)
+def on_migrate(sender, **kwargs):
+    from .models import SupplierModule
+
+    SupplierModule.ensure_all_supplier_modules()
 
 
 def handle_post_save_bump_all_prices_caches(sender, instance, **kwargs):
@@ -87,7 +94,7 @@ def on_order_creator_finished(sender, order, source, **kwargs):
 @receiver(order_changed)
 def on_order_changed(sender, order, **kwargs):
     for line in order.lines.products().only("product_id", "supplier").select_related("supplier"):
-        line.supplier.module.update_stock(line.product_id)
+        line.supplier.update_stock(line.product_id)
 
 
 def handle_contact_group_price_display_post_save(sender, instance, **kwargs):