+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Changed
+
+- Admin: filter products in list and edit views according to product kind listing name
+- Core: suppliers don't have default supplier modules anymore
+
+### Add
+
+- Core: Allow a supplier to have multiple modules
+
 ### Fixed
 
 - Notify: always overrride the current notification data with the new one